+++ src/Achievements/src/Achievement.h
@@ -1,0 +1,13 @@
+//
+// Created by remy on 13-07-20.
+//
+
+#ifndef C_OOKIECLIKER_ACHIEVEMENT_H
+#define C_OOKIECLIKER_ACHIEVEMENT_H
+class Achievement {
+public:
+
+private:
+
+};
+#endif //C_OOKIECLIKER_ACHIEVEMENT_H