+++ src/Achievements/src/Achievements.h
@@ -1,0 +1,17 @@
+//
+// Created by remy on 21-04-20.
+//
+
+#ifndef C_OOKIECLIKER_Achievements_H
+#define C_OOKIECLIKER_Achievements_H
+#include "CookieNumbers.h"
+
+class Achievements {
+    public:
+
+private:
+
+};
+
+
+#endif